requirements.txt
Flask==3.1.2
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1761567966/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1764766840/work
joblib==1.5.2
matplotlib==3.10.7
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mysql-connector-python==9.5.0
numpy==2.3.5
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.16.3
